// services/gmail/pod-gmail-next/src/message/types.ts
//
// Copyright © 2025 Hardcore Engineering Inc.
//
// Licensed under the Eclipse Public License, Version 2.0 (the 'License');
// you may not use this file except in compliance with the License. You may
// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an 'AS IS' BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//
// See the License for the specific language governing permissions and
// limitations under the License.
//

export interface AttachedFile {
  id: string
  name: string
  data: Buffer
  contentType: string
  size?: number
  lastModified: number
}

export interface EmailContact {
  email: string
  firstName?: string
  lastName?: string
  photoUrl?: string | null
}